require authentication

+++ public/lib/delete_file.php
@@ -1,9 +1,15 @@
 <?php 
+	session_start();
+	if(!isset($_SESSION['uid'])) {
+		// TODO: proper error page
+		header('Location: '. '../index.php');
+		die("Not authenticated");
+	}
 	include_once("Connection.php");
 
 	if(isset($_POST['id'])) {
 		$c = Connection::Instance();
-		$id = $_POST['id'];
+		$id = (int) $_POST['id'];
 		$sql = sprintf("DELETE FROM File WHERE _id = %s;", $id);
 		try {
 			$statment = $c->connection->query($sql);
